Refactor: Improve code quality in mobile editor components
- Fix type safety by properly typing HotspotSize in TabContent
- Move MOBILE_COLORS and SIZE_OPTIONS constants to TabContent for better encapsulation
- Enhance accessibility with aria-label for back button
- Remove unused MOBILE_COLORS prop from MobileEditorModal
